Added new optional parameters 'port' and 'req_protocols'.
 * testenv/test/base_test.py (BaseTest.__init__): Added optional
   parameters 'port' and 'req_protocols'.
   (gen_cmd_line): Construct request URIs from list 'req_protocol'.
 * testenv/test/http_test.py (HttpTest.__init__): Added optional
   parameters 'port' and 'req_protocols'.
   (instantiate_server_by): Listen address constructed according
   to port.

The ability for tests to be able to explicitly set some uncommon parameters
such as the port number or the actual requested protocol became obvious when
support for HSTS was added to Wget, where the requested URI and the actual executed
URI do not always have to be the same. These parameters are optional, and if they're
not specified, the test suite behaves as before. This way we preserve backward-compatibility.